feat: Require list of expected algorithms when secret/publicKey is given
Instead of allowing only one single algorithm during signature validation,
one can now specify a comma-separated list of algorithms using the
`--algs` command line parameter.

In order to encourage users to be aware of the choice of algorithms and
safely define a subset of the supported algorithms, the `--algs`
parameter is now required when the `-S` parameter is set.